mp-mega: split felt and belt mpeval
In the event that a gate is pinned to the subject that is not part of normal evaluation via cenhep or similar, the jets for that gate will never fire. This causes verification to take several minutes, regardless of of the number of jets in the miner. This is likely causing connections to nodes to timeout during verification, aggravating networking issues.

It appears that it does not work to use tistar instead, as all dynamic
gate lookups appear to be broken.